Remote/VM/Docker: make them optional plugins
Our users have pointed out that the basic Avocado test runner, and its
related packaged form, has too many dependencies.  They're absolutely
right, and the whole goal of having plugins in Avocado is that
dependencies can be made optional according to the features required
by a given user.

This specific change creates three new optional plugins, containing
the extended test runners: remote, vm and docker.
