--- avocado/core/remote/__init__.py
@@ -1,17 +1,0 @@
-# This program is free software; you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation; either version 2 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-#
-# See LICENSE for more details.
-#
-# Copyright: Red Hat Inc. 2014-2015
-# Author: Ruda Moura <rmoura@redhat.com>
-
-from .runner import RemoteTestRunner, VMTestRunner
-
-__all__ = ['RemoteTestRunner', 'VMTestRunner']
