+++ optional_plugins/runner_docker/avocado_runner_docker/__init__.py
@@ -19,17 +19,19 @@ import socket
 import time
 
 import aexpect
+
 from avocado.core.plugin_interfaces import CLI
-from avocado.core.remote.runner import RemoteTestRunner
 from avocado.utils import process
 from avocado.utils.wait import wait_for
+from avocado_runner_remote import RemoteTestRunner
 
 
 class DockerRemoter(object):
 
     """
-    Remoter object similar to `avocado.core.remoter` which implements subset
-    of the commands on docker container.
+    Remoter object similar to
+    :class:`avocado_runner_remoter.RemoteTestRunner` which implements
+    subset of the commands on docker container.
     """
 
     def __init__(self, dkrcmd, image, options, name=None):
@@ -140,7 +142,7 @@ class DockerTestRunner(RemoteTestRunner):
                 self.job.log.warn("DOCKER     : Fail to cleanup: %s" % details)
 
 
-class Docker(CLI):
+class DockerCLI(CLI):
 
     """
     Run the job inside a docker container
