+++ selftests/unit/test_remote.py
@@ -1,16 +1,14 @@
 #!/usr/bin/env python
 
-import unittest
+import logging
 import os
+import unittest
 
 from flexmock import flexmock, flexmock_teardown
 
-from avocado.core import remoter
-from avocado.core import remote
 from avocado.utils import archive
-import logging
+import avocado_runner_remote
 
-cwd = os.getcwd()
 
 JSON_RESULTS = ('Something other than json\n'
                 '{"tests": [{"test": "1-sleeptest;0",'
@@ -52,8 +50,8 @@ class RemoteTestRunnerTest(unittest.TestCase):
                        logdir="/local/path",
                        _result_events_dispatcher=result_dispatcher)
 
-        flexmock(remote.RemoteTestRunner).should_receive('__init__')
-        self.runner = remote.RemoteTestRunner(job, None)
+        flexmock(avocado_runner_remote.RemoteTestRunner).should_receive('__init__')
+        self.runner = avocado_runner_remote.RemoteTestRunner(job, None)
         self.runner.job = job
 
         filehandler = logging.StreamHandler()
@@ -63,7 +61,7 @@ class RemoteTestRunnerTest(unittest.TestCase):
         stream = flexmock(job_unique_id='1-sleeptest;0',
                           debuglog='/local/path/dirname')
         Remote = flexmock()
-        Remoter = flexmock(remoter.Remote)
+        Remoter = flexmock(avocado_runner_remote.Remote)
         Remoter.new_instances(Remote)
         args_version = 'avocado -v'
         version_result = flexmock(stdout='Avocado 1.2', exit_status=0)
@@ -148,7 +146,7 @@ class RemoteTestRunnerSetup(unittest.TestCase):
 
     def setUp(self):
         Remote = flexmock()
-        remote_remote = flexmock(remoter)
+        remote_remote = flexmock(avocado_runner_remote)
         (remote_remote.should_receive('Remote')
          .with_args(hostname='hostname', username='username',
                     password='password', key_filename=None, port=22,
@@ -169,7 +167,7 @@ class RemoteTestRunnerSetup(unittest.TestCase):
         log = flexmock()
         log.should_receive("info")
         job = flexmock(args=Args, log=log)
-        self.runner = remote.RemoteTestRunner(job, None)
+        self.runner = avocado_runner_remote.RemoteTestRunner(job, None)
 
     def tearDown(self):
         flexmock_teardown()
