+++ setup.py
@@ -134,11 +134,8 @@ if __name__ == '__main__':
                   'xunit = avocado.plugins.xunit:XUnitCLI',
                   'json = avocado.plugins.jsonresult:JSONCLI',
                   'journal = avocado.plugins.journal:Journal',
-                  'remote = avocado.plugins.remote:Remote',
                   'replay = avocado.plugins.replay:Replay',
                   'tap = avocado.plugins.tap:TAP',
-                  'vm = avocado.plugins.vm:VM',
-                  'docker = avocado.plugins.docker:Docker',
                   'yaml_to_mux = avocado.plugins.yaml_to_mux:YamlToMux',
                   'zip_archive = avocado.plugins.archive:ArchiveCLI',
                   ],
